[package] udev: bump to 173
Yet another update, I inherited a wrong behaviour with older versions.
It used to $(INSTALL_BIN) symlinks which results in copying the dereferenced
file and not the symlink itself, now uses $(CP) and it's reported working (and running)
by at least one person in IRC (using a glibc toolchain with mips). Works and runs on
my toolchain (mipsel uClibc) too.

Signed-off-by: Daniel Mierswa <impulze@impulze.org>