[package] ffmpeg: Allow for more configurable FFmpeg build options (#7837, #8465)
This patch allows broad control over FFmpeg's libav* build configuration
and fixes a typo in ffprobe's install section.

The patch resolves #7837 as the choice of building a full libav* is available.
This should also resolve #8465 as input devices such as cameras can be enabled.

The files created by a full libav* build will be large (several megabytes) and
not recommended for end users. Its use could help debugging other issues to
learn whether the problem is the openwrt build configuration or another cause.

Signed-off-by: Ian Leonard <antonlacon@gmail.com>