au1000: fix InternetBox idsel and remove broken flag
The following patch is necessary for the InternetBox (Surfbox) board-specific
idsel to work with newer kernels. Acutally I don't exactly understand why this
was not necessary before, but without the udelay, the MiniPCI adapter in slot 3
is not detected and set up correctly.

Also remove platform broken flag, it works.

Signed-off-by: Bruno Randolf <br1@einfach.org>