ar71xx: create 'fat' images for the DIR-825/TEW-673GRU boards
With these images, it is possible to use the 'unused'
partition of the flash. The 'fat' images can be installed
with the sysupgrade command. When a 'fat' image is
installed from a regular one, the platform specific
sysupgrade script copies the calibration data to the
end of the flash. Likewise, when a regular image is
installed  from the 'fat' version the sysupgrade script
copies the calibration data back to the original location.