Fix ARV4518PW wifi issue. AA/12.09-rc2.
This patch corrects an issue with the wifi for the ARV4518PW in Attitude Adjustment that involves the PCI bus configuration in this board. The bootlog shows: ath5k: phy0: failed to warm reset the MAC Chip.
Commenting this line forces the board definition to be configured with the default PCI_GNT1 | PCI_REQ1 defined in line 138.
The wifi and the usb have been deeply tested and they worked without problems. Maybe would be a good idea to make stress tests.
Once applied this ticket should be closed:
https://dev.openwrt.org/ticket/11704

Signed-off-by: José Vázquez Fernández <ppvazquezfer@gmail.com>
Signed-off-by: Daniel González <dgcbueu@gmail.com>