ar71xx: image: add new macros for AP99 based Cameo devices
The MTD layout of the AP99 based devices is different.
Add new macros to reflect that and use the new macros
for the DIR-615-E4 board.

This prevents overwriting of the 'mac' partition which
contain vital informations. The disadvantage of the change
is that the usable space is reduced by 192KiB. Blame the
vendors for the unfriendly partition layout.

Signed-off-by: Gabor Juhos <juhosg@openwrt.org>