busybox: fix existing "export -n" in ash
Minor fix to existing patch which did not correctly implement -n support for the export command in ash. The VEXPORT flag of a variable was set in the first call to setvareq() and then remained set in the subsequent call, because the variable already existed in the environment.

Although /lib/functions.sh by default does not export variables (NO_EXPORT=1), this issue was leading to dirty environments.

Without this fix:
(ash) export A=1; export -n B=2; /bin/ash -c "echo \$A, \$B"
1, 2

With this fix:
(ash) export A=1; export -n B=2; /bin/ash -c "echo \$A, \$B"
1,

Signed-off-by: Christophe Lermytte <openwrt@lermytte.be>