ramips: Add support for the 7Links PX-4885 wireless router.
The 7Links PX-4885 is a small battery-powered wireless router.
It is based on a RT5350F WiSoC and features one ethernet port and one
USB port. It is a Hame MPR-A1 clone, except it has different GPIOs,
a different battery and more RAM (32 MB).
This patch adds a profile for this board, the corresponding device tree
file and the adequate base-files changes (incl. sysupgrade).

Signed-off-by: Hugo Grostabussiat <bonstra@bonstra.fr.eu.org>