include: fix kernel download failure with git & local trees
When using the options EXTERNAL_KERNEL_TREE or KERNEL_GIT_CLONE_URI,
the command "make downloads" fails as it tries to download the kernel
tarball despite the option. This doesn't happen during a regular build
as in that case, the dependency is conditionned through the LINUX_SITE
variable, which is not set in these cases.

Below is a snapshot of the error for an target using a 3.14 kernel:

make[3]: *** No rule to make target `.../dl/linux-3.14.tar.xz',
needed by `download'.  Stop.

Change-Id: I1244969c1bbf9c81a6a64d68ae88ac58b0f8e79e
Signed-off-by: Mathieu Olivari <mathieu@qca.qualcomm.com>